requirements.txt
openai==1.35.3
python-dotenv
azure-search-documents



